Module Ambient_span_provider.Callbacks

type 'st t = {

with_current_span_set_to : 'a. 'st -> Trace_core__.Types.span -> (Trace_core__.Types.span -> 'a) -> 'a;
(*
with_current_span_set_to span f sets the span as current span, enters f span, and restores the previous current span if any
*)

get_current_span : 'st -> Trace_core__.Types.span option;
(*
Access the current span from some ambient scope. This is only supported for collectors that provide a current_span_wrap field.
*)

}