Module Trace_core.Collector

A global collector.

The collector, if present, is responsible for collecting messages and spans, and storing them, recording them, forward them, or offering them to other services and processes.

type Trace_core__.Types.span +=

| Span_dummy

val dummy_span : Trace_core__.Types.span

A fake span that never emits data. All collectors should handle this span by doing nothing.

module Callbacks : sig ... end

type t =

| C_none
(*
No collector.
*)

| C_some : 'st * 'st Callbacks.t -> t
(*
Collector with a state and some callbacks.
*)

Definition of a collector.

This is only relevant to implementors of tracing backends; to instrument your code you only need to look at the Trace module.

The definition changed since 0.11 to a record of callbacks + a state

val is_some : t -> bool