Module type Collector.S

Signature for a collector.

This is only relevant to implementors of tracing backends; to instrument your code you only need to look at the Trace module.

val enter_span : ?__FUNCTION__:string -> __FILE__:string -> __LINE__:int -> data: (string * [ `Int of int | `String of string | `Bool of bool | `None ]) list -> string -> int64

val exit_span : int64 -> unit

val message : ?span:int64 -> data: (string * [ `Int of int | `String of string | `Bool of bool | `None ]) list -> string -> unit

val name_thread : string -> unit

Give a name to the current thread

val name_process : string -> unit

Give a name to the current process

val counter_int : string -> int -> unit

Integer counter

val counter_float : string -> float -> unit

Float counter

val shutdown : unit -> unit

Shutdown collector, possibly waiting for it to finish sending data.