Module Trace_subscriber

Generic subscribers.

This defines the notion of a subscriber, a set of callbacks for every trace event. It also defines a collector that needs to be installed for the subscriber(s) to be called.

module Callbacks : sig ... end

Callbacks used for subscribers.

module Subscriber : sig ... end

Trace subscribers

type user_data =

| U_bool of bool

| U_float of float

| U_int of int

| U_none

| U_string of string
(*
A non polymorphic-variant version of Trace_core.user_data
*)

type flavor =

| Sync

| Async
(*
A non polymorphic-variant version of Trace_core.flavor
*)

Main API

type t = Subscriber.t

val collector : t -> Trace_core.collector

A collector that calls the subscriber's callbacks.

It uses mtime (if available) to obtain timestamps.