Module Trace_util.Multi_collector

Combine multiple collectors into one

val combine_l : Trace_core.Collector.t list -> Trace_core.Collector.t

Combine multiple collectors, ie return a collector that forwards to every collector in the list.

val combine : Trace_core.Collector.t -> Trace_core.Collector.t -> Trace_core.Collector.t

combine s1 s2 is a collector that forwards every call to s1 and s2 both.