Module Trace

Trace.

type span = int64

A span identifier.

type user_data = [

| `Int of int

| `String of string

| `Bool of bool

| `None

]

User defined data, generally passed as key/value pairs

module Collector : sig ... end

A global collector.

Tracing

val enabled : unit -> bool

Is there a collector?

This is fast, so that the traced program can check it before creating any span or message

val enter_span : ?__FUNCTION__:string -> __FILE__:string -> __LINE__:int -> ?data:(unit -> (string * user_data) list) -> string -> span

val exit_span : span -> unit

val with_span : ?__FUNCTION__:string -> __FILE__:string -> __LINE__:int -> ?data:(unit -> (string * user_data) list) -> string -> (span -> 'a) -> 'a

val message : ?span:span -> ?data:(unit -> (string * user_data) list) -> string -> unit

val messagef : ?span:span -> ?data:(unit -> (string * user_data) list) -> ((('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a) -> unit) -> unit

val set_thread_name : string -> unit

Give a name to the current thread.

val set_process_name : string -> unit

Give a name to the current process.

val counter_int : string -> int -> unit

Emit a counter (int)

val counter_float : string -> float -> unit

Emit a counter (float)

Collector

type collector = (module Collector.S)

val setup_collector : collector -> unit

setup_collector c installs c as the collector.

raises Invalid_argument
if there already is an established collector.

val shutdown : unit -> unit