Module Trace_core.Collector

A global collector.

The collector, if present, is responsible for collecting messages and spans, and storing them, recording them, forward them, or offering them to other services and processes.

val dummy_span : int64

val dummy_explicit_span : Trace_core__.Types.explicit_span

module type S = sig ... end

Signature for a collector.