Module Callbacks.Dummy

Dummy callbacks. It can be useful to reuse some of these functions in a real subscriber that doesn't want to handle all events, but only some of them.

To write a subscriber that only supports some callbacks, this can be handy:

  module My_callbacks = struct
  type st = my_own_state
  include Callbacks.Dummy
  let on_counter (st:st) ~time_ns ~tid ~data ~name v : unit = ...
  end

val on_init : 'a -> time_ns:'b -> unit

val on_shutdown : 'a -> time_ns:'b -> unit

val on_name_thread : 'a -> time_ns:'b -> tid:'c -> name:'d -> unit

val on_name_process : 'a -> time_ns:'b -> tid:'c -> name:'d -> unit

val on_message : 'a -> time_ns:'b -> tid:'c -> span:'d -> data:'e -> 'f -> unit

val on_counter : 'a -> time_ns:'b -> tid:'c -> data:'d -> name:'e -> 'f -> unit

val on_enter_span : 'a -> __FUNCTION__:'b -> __FILE__:'c -> __LINE__:'d -> time_ns:'e -> tid:'f -> data:'g -> name:'h -> 'i -> unit

val on_exit_span : 'a -> time_ns:'b -> tid:'c -> 'd -> unit

val on_add_data : 'a -> data:'b -> 'c -> unit

val on_enter_manual_span : 'a -> __FUNCTION__:'b -> __FILE__:'c -> __LINE__:'d -> time_ns:'e -> tid:'f -> parent:'g -> data:'h -> name:'i -> flavor:'j -> trace_id:'k -> 'l -> unit

val on_exit_manual_span : 'a -> time_ns:'b -> tid:'c -> name:'d -> data:'e -> flavor:'f -> trace_id:'g -> 'h -> unit

val on_extension_event : 'a -> time_ns:'b -> tid:'c -> 'd -> unit